[refactor] Split key sagas.
